handle disconnects/connects/reconnects better.
before this commit, during a (re)connect, all styx operations were
halted.  on timeout of the dial(), ircfs seemed to hang.  now we
keep the styx write that requested the reconnect, and keep handling
new styx ops.  before exec'ing the op, we check if the connection
state is valid for the op (connected normally).  as soon as the
dial/irc logon fails, we respond on the ctl write.  so an "echo
reconnect >ctl" succeeding still means we are connected.

also, make ircfs clear itself on exit (due to styx eof).