try a bit harder not to send irc messages to the server that are too long.
after login or a nick change, whois asks for a whois on ourself.
when the response comes in, it will have the user & host name used.
we use that to calculate what the server prepends to our messages
before they are sent to others.  for most irc servers, this will
be accurate.

as requested by soul9 on #inferno.