don't block when writing to a window's ctl file.
when the connection to ircfs is hanging, or on a "connect" or
"reconnect" ctl message, wm/irc no longer "hangs" (wait for the
write on the ctl file to return).  this also removes the per-window
writer process, but instead always opens the ctl file for each
window.

this commit also fixes the datatype of a line of text that's the
result of a /msg command.