do not crash on wstat on not-open data files.
found by fr1tz on #plan9.  9pfuse/linux 9p will do a wstat of lenght=0
when doing a ">" redirection in the shell, before the file is opened.
that was (incorrectly) not expected.

note that using ircfs with 9pfuse/linux 9p from the shell is still
suboptimal (at least on my linux machine): opens are done with O_APPEND,
or with with other redirection it is not always easy to open only for
writing without truncating/appending.

this also removes -D from the usage line.